Base packages in defaults
Move all packages and services mentioned in the base_packages role to
defaults/ so they can be overridden on a per-host basis.

Moved services to disable before removing packages so any that may be
removed are not left in an odd state when removed (had that happen with
at least one service).

Added some optional "additional" variables so you can add packages to
the existing defaults without having to redefine the entire list.